requirements.txt
WTForms==2.2.1
SQLAlchemy==1.2.7
bcrypt==3.1.7
requests==2.18.4
flask-heroku==0.1.9
gunicorn==19.9.0
keras==2.3.1
tensorflow==2.0.0
tensorflow-estimator==2.0.0
Pillow==5.1.0
# Flask-Images==3.0.2
stripe==2.18.1
# stripe==2.37.2
# stripe
